ISSUE #5412 updated calls to generate sessions
