+++ backend/tests/v5/e2e/services/chat/index.test.js
@@ -133,8 +133,7 @@ const testUnauthenticatedUser = () => {
 
 		test('should be able to join rooms after login', async () => {
 			const socket = await ServiceHelper.socket.connectToSocket();
-			await ServiceHelper.loginAndGetCookie(agent, tsAdmin.user, tsAdmin.password,
-				{ [SOCKET_HEADER]: socket.id });
+			await ServiceHelper.loginAndGetCookie(agent, tsAdmin, { [SOCKET_HEADER]: socket.id });
 
 			// introduce a delay to let backend sort out their events (i.e. hooking the session with the socket)
 			await ServiceHelper.sleepMS(100);
@@ -148,7 +147,7 @@ const testTSAdmin = () => {
 	describe('User with access', () => {
 		let cookie;
 		beforeAll(async () => {
-			const cookieInfo = await ServiceHelper.loginAndGetCookie(agent, tsAdmin.user, tsAdmin.password);
+			const cookieInfo = await ServiceHelper.loginAndGetCookie(agent, tsAdmin);
 			cookie = cookieInfo.session;
 		});
 		test('should be able to connect to the chat service', async () => {
@@ -207,7 +206,7 @@ const testNobody = () => {
 	describe('User without a license', () => {
 		let cookie;
 		beforeAll(async () => {
-			const cookieInfo = await ServiceHelper.loginAndGetCookie(agent, nobody.user, nobody.password);
+			const cookieInfo = await ServiceHelper.loginAndGetCookie(agent, nobody);
 			cookie = cookieInfo.session;
 		});
 		test('should be able to connect to the chat service', async () => {
